requirements/base.txt
psycopg2==2.6.1
Django==1.8.5
djangorestframework==3.3.1